### Summary

Update internal dogfood branch to latest NCS.

### Test Plan

Performed the merge locally and confirmed
successful build with asset tracker app.

+++ samples/cellular/modem_shell/src/link/link.c
@@ -34,8 +34,6 @@
 #include "mosh_defines.h"
 
 #include <nrf_modem_at.h>
-#include <modem/at_cmd_parser.h>
-#include <modem/at_params.h>
 
 extern bool uart_shell_disable_during_sleep_requested;
 extern struct k_work_q mosh_common_work_q;
@@ -443,7 +441,7 @@ void link_ind_handler(const struct lte_lc_evt *const evt)
 		len = snprintf(
 			log_buf, sizeof(log_buf),
 			"eDRX parameter update: eDRX: %f, PTW: %f",
-			evt->edrx_cfg.edrx, evt->edrx_cfg.ptw);
+			(double)evt->edrx_cfg.edrx, (double)evt->edrx_cfg.ptw);
 		if (len > 0) {
 			mosh_print("%s", log_buf);
 		}
